requirements.txt
numpy==1.21.2
matplotlib==3.2.0
scipy==1.7.1
pandas==1.3.3
botocore==1.21.13
ibm_cos_sdk==2.10.0
scikit_learn==1.0.1
=======
streamlit==1.2.0
streamlit-player==0.1.5
requests==2.26.0
python-decouple==3.5

